fix: merged dash with availability bar
