+++ lib/deepotsava/accounting/details.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:garuda/deepotsava/datatypes.dart';
+import 'package:garuda/deepotsava/sales/datatypes.dart';
 import 'package:garuda/deepotsava/fbl.dart';
 
 class Details extends StatefulWidget {
