+++ lib/deepotsava/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:garuda/deepotsava/accounting/main.dart';
-import 'package:garuda/deepotsava/sales.dart';
+import 'package:garuda/deepotsava/sales/main.dart';
 import 'package:garuda/theme.dart';
 import 'package:garuda/toaster.dart';
