+++ lib/deepotsava/sales/log.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:garuda/deepotsava/datatypes.dart';
+import 'package:garuda/deepotsava/sales/datatypes.dart';
 import 'package:garuda/deepotsava/fbl.dart';
-import 'package:garuda/deepotsava/log_dialog.dart';
+import 'package:garuda/deepotsava/sales/log_dialog.dart';
 import 'package:garuda/theme.dart';
 
 class Log extends StatefulWidget {
