+++ lib/deepotsava/sales/stock_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:garuda/deepotsava/datatypes.dart';
+import 'package:garuda/deepotsava/sales/datatypes.dart';
 import 'package:garuda/deepotsava/fbl.dart';
-import 'package:garuda/deepotsava/stock_dialog.dart';
+import 'package:garuda/deepotsava/sales/stock_dialog.dart';
 import 'package:garuda/theme.dart';
 import 'package:intl/intl.dart';
